requirements.txt
python-dotenv
pydantic
openai
# openai-agents
# openai-agent[voice]